requirements.txt
pytest==8.3.3
pytest-html==4.1.1
PyYAML==6.0.2
pyodbc==4.0.34
pymssql==2.3.1